src/test/java/com/atopybook/core/app/service/MemberServiceTest.java
package com.atopybook.core.app.service;

import org.assertj.core.api.Assertions;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.transaction.annotation.Transactional;

import static org.assertj.core.api.Assertions.assertThat;
import static org.junit.jupiter.api.Assertions.*;

@SpringBootTest
@Transactional
class MemberServiceTest {

    @Autowired
    MemberService memberService;

    @Test
    public void join() {
        String memberName = "jayden";

        Long memberId = memberService.join(memberName);
        assertThat(memberId).isEqualTo(1L);
    }
}